move blas routines to blas.cpp
